Build iso2022 as a separate unit.
This is for MSVC to trim the unused conversion table.
I feel lame, but /GL+/LTCG doesn't help.

If I understand correctly, custom UTF-8 handling of MultiByteToWideChar
and the opposite one (splitted out) are for Windows 9x though.

It may require corresponding dependency update for non-Windows,
but it is untouched.